Adds a check to make sure the pages table exisits before trying to drop it.

+++ db/migrate/20171216191153_remove_pages_table.rb
@@ -1,5 +1,5 @@
 class RemovePagesTable < ActiveRecord::Migration[5.1]
   def change
-    drop_table :pages
+    drop_table :pages if ActiveRecord::Base.connection.table_exists?  'pages'
   end
 end
